Fix: Se ha estructurizado el proyecto, el cual se ha explicado en el README
