+++ frontend/src/styles/Lineas.css
