+++ frontend/src/styles/output.css
