+++ frontend/tailwind.cmd
@@ -1,0 +1,1 @@
+npx @tailwindcss/cli -i ./src/styles/input.css -o ./src/styles/output.css --watch
