--- tailwind.cmd
@@ -1,1 +1,0 @@
-npx @tailwindcss/cli -i ./src/input.css -o ./src/output.css --watch
